+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - PATCH version when backwards compatible bug **fixes** are implemented.
 
 ## [Unreleased]
+### Added
+- merchantSession, merchantCard, merchantInstallment and merchantPurchase resources
 
 ## [2.25.0] - 2026-01-07
 ### Added